+++ bspline_fitting/Cpp/src/value.cpp
@@ -1,11 +1,11 @@
 #include "value.h"
 #include <cmath>
 
-const std::vector<float> linspace(const float &start, const float &stop,
-                                  const int &num) {
-  std::vector<float> ret_vec(1);
+const std::vector<double> linspace(const double &start, const double &stop,
+                                   const int &num) {
+  std::vector<double> ret_vec(1);
 
-  const float delta = stop - start;
+  const double delta = stop - start;
 
   if (std::fabs(delta) <= 10e-8) {
     ret_vec[0] = start;
@@ -22,7 +22,7 @@ const std::vector<float> linspace(const float &start, const float &stop,
   const int div = num - 1;
 
   for (int i = 0; i < num; ++i) {
-    const float current_value = start + i * delta / div;
+    const double current_value = start + i * delta / div;
     ret_vec[i] = current_value;
   }
 
@@ -30,8 +30,8 @@ const std::vector<float> linspace(const float &start, const float &stop,
 }
 
 const int find_span_linear(const int &degree,
-                           const std::vector<float> &knot_vector,
-                           const int &num_ctrlpts, const float &knot) {
+                           const std::vector<double> &knot_vector,
+                           const int &num_ctrlpts, const double &knot) {
   int span = degree + 1;
 
   while (span < num_ctrlpts && knot_vector[span] <= knot) {
@@ -42,13 +42,13 @@ const int find_span_linear(const int &degree,
 }
 
 const std::vector<int> find_spans(const int &degree,
-                                  const std::vector<float> &knot_vector,
+                                  const std::vector<double> &knot_vector,
                                   const int &num_ctrlpts,
-                                  const std::vector<float> &knots) {
+                                  const std::vector<double> &knots) {
   std::vector<int> spans(knots.size());
 
   for (size_t i = 0; i < knots.size(); ++i) {
-    const float &knot = knots[i];
+    const double &knot = knots[i];
 
     const int current_span =
         find_span_linear(degree, knot_vector, num_ctrlpts, knot);
@@ -59,21 +59,21 @@ const std::vector<int> find_spans(const int &degree,
   return spans;
 }
 
-const std::vector<float> basis_function(const int &degree,
-                                        const std::vector<float> &knot_vector,
-                                        const int &span, const float &knot) {
-  std::vector<float> left(degree + 1, 0.0);
-  std::vector<float> right(degree + 1, 0.0);
-  std::vector<float> N(degree + 1, 1.0);
+const std::vector<double> basis_function(const int &degree,
+                                         const std::vector<double> &knot_vector,
+                                         const int &span, const double &knot) {
+  std::vector<double> left(degree + 1, 0.0);
+  std::vector<double> right(degree + 1, 0.0);
+  std::vector<double> N(degree + 1, 1.0);
 
   for (int j = 1; j < degree + 1; ++j) {
     left[j] = knot - knot_vector[span + 1 - j];
     right[j] = knot_vector[span + j] - knot;
 
-    float saved = 0.0;
+    double saved = 0.0;
 
     for (int r = 0; r < j; ++r) {
-      const float temp = N[r] / (right[r + 1] + left[j - r]);
+      const double temp = N[r] / (right[r + 1] + left[j - r]);
       N[r] = saved + right[r + 1] * temp;
       saved = left[j - r] * temp;
     }
@@ -84,17 +84,17 @@ const std::vector<float> basis_function(const int &degree,
   return N;
 }
 
-const std::vector<std::vector<float>>
-basis_functions(const int &degree, const std::vector<float> &knot_vector,
+const std::vector<std::vector<double>>
+basis_functions(const int &degree, const std::vector<double> &knot_vector,
                 const std::vector<int> &spans,
-                const std::vector<float> &knots) {
-  std::vector<std::vector<float>> basis(spans.size());
+                const std::vector<double> &knots) {
+  std::vector<std::vector<double>> basis(spans.size());
 
   for (size_t i = 0; i < spans.size(); ++i) {
     const int &span = spans[i];
-    const float &knot = knots[i];
+    const double &knot = knots[i];
 
-    const std::vector<float> current_basis_function =
+    const std::vector<double> current_basis_function =
         basis_function(degree, knot_vector, span, knot);
 
     basis[i] = current_basis_function;
@@ -103,27 +103,27 @@ basis_functions(const int &degree, const std::vector<float> &knot_vector,
   return basis;
 }
 
-const std::vector<std::vector<float>>
+const std::vector<std::vector<double>>
 toPoints(const int &degree_u, const int &degree_v, const int &size_u,
          const int &size_v, const int &sample_num_u, const int &sample_num_v,
-         const float &start_u, const float &start_v, const float &stop_u,
-         const float &stop_v, const std::vector<float> &knotvector_u,
-         const std::vector<float> &knotvector_v,
-         const std::vector<std::vector<float>> &ctrlpts) {
-  const std::vector<float> knots_u = linspace(start_u, stop_u, sample_num_u);
-  const std::vector<float> knots_v = linspace(start_v, stop_v, sample_num_v);
+         const double &start_u, const double &start_v, const double &stop_u,
+         const double &stop_v, const std::vector<double> &knotvector_u,
+         const std::vector<double> &knotvector_v,
+         const std::vector<std::vector<double>> &ctrlpts) {
+  const std::vector<double> knots_u = linspace(start_u, stop_u, sample_num_u);
+  const std::vector<double> knots_v = linspace(start_v, stop_v, sample_num_v);
 
   const std::vector<int> spans_u =
       find_spans(degree_u, knotvector_u, size_u, knots_u);
   const std::vector<int> spans_v =
       find_spans(degree_v, knotvector_v, size_v, knots_v);
 
-  const std::vector<std::vector<float>> basis_u =
+  const std::vector<std::vector<double>> basis_u =
       basis_functions(degree_u, knotvector_u, spans_u, knots_u);
-  const std::vector<std::vector<float>> basis_v =
+  const std::vector<std::vector<double>> basis_v =
       basis_functions(degree_v, knotvector_v, spans_v, knots_v);
 
-  std::vector<std::vector<float>> eval_points;
+  std::vector<std::vector<double>> eval_points;
 
   for (size_t i = 0; i < spans_u.size(); ++i) {
     const int idx_u = spans_u[i] - degree_u;
@@ -131,14 +131,14 @@ toPoints(const int &degree_u, const int &degree_v, const int &size_u,
     for (size_t j = 0; j < spans_v.size(); ++j) {
       const int idx_v = spans_v[j] - degree_v;
 
-      std::vector<float> spt(3, 0.0);
+      std::vector<double> spt(3, 0.0);
 
       for (int k = 0; k < degree_u + 1; ++k) {
-        std::vector<float> temp(3, 0.0);
+        std::vector<double> temp(3, 0.0);
 
         for (int l = 0; l < degree_v + 1; ++l) {
           for (size_t m = 0; m < temp.size(); ++m) {
-            const float cp = ctrlpts[idx_v + l + (size_v * (idx_u + k))][m];
+            const double cp = ctrlpts[idx_v + l + (size_v * (idx_u + k))][m];
 
             temp[m] += basis_v[j][l] * cp;
           }
